tidied up firebase imports and turned off swift whole module optimisation in debug
